bug and pre-commit fixes
