+++ .idea/workspace.xml
@@ -6,17 +6,8 @@
   <component name="ChangeListManager">
     <list default="true" id="7da7499e-1ec6-4bd7-974b-80f0bb5c6fdd" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/flask_app/broker/name.py" beforeDir="false" afterPath="$PROJECT_DIR$/flask_app/broker/name.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/sppy/tools/provider/api.py" beforeDir="false" afterPath="$PROJECT_DIR$/sppy/tools/provider/api.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/sppy/tools/provider/gbif.py" beforeDir="false" afterPath="$PROJECT_DIR$/sppy/tools/provider/gbif.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/sppy/tools/provider/idigbio.py" beforeDir="false" afterPath="$PROJECT_DIR$/sppy/tools/provider/idigbio.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/sppy/tools/provider/ipni.py" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/sppy/tools/provider/itis.py" beforeDir="false" afterPath="$PROJECT_DIR$/sppy/tools/provider/itis.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/sppy/tools/provider/mopho.py" beforeDir="false" afterPath="$PROJECT_DIR$/sppy/tools/provider/mopho.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/sppy/tools/provider/worms.py" beforeDir="false" afterPath="$PROJECT_DIR$/sppy/tools/provider/worms.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/sppy/tools/s2n/lm_xml.py" beforeDir="false" afterPath="$PROJECT_DIR$/sppy/tools/s2n/lm_xml.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/sppy/tools/s2n/solr.py" beforeDir="false" afterPath="$PROJECT_DIR$/sppy/tools/s2n/solr.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/sppy/tools/s2n/utils.py" beforeDir="false" afterPath="$PROJECT_DIR$/sppy/tools/s2n/utils.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/sppy/tools/fileop/ready_file.py" beforeDir="false" afterPath="$PROJECT_DIR$/sppy/tools/fileop/ready_file.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/sppy/tools/fileop/split_providers.py" beforeDir="false" afterPath="$PROJECT_DIR$/sppy/tools/fileop/split_records.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
